Protovis values and formatting correct (mostly)
I have to replace the '_' with ' ' for the indent, but other than that
looks brilliant!

Ready to test with actual protovis html stuff. :D

+++ OronymCode/2dStackedBarTreeViz.h
@@ -13,7 +13,7 @@ void drawStackedBar( int width, std::vector <std::string> tailPhrases);
 std::string removeSuccessAndDeadendIndicatorsFromString( std::string orig );
 std::vector <std::string> stripOronymOutputOfEndIndicators ( std::vector <std::string>  origPhrases ) ;
 void makeProtvisDiagram (std::vector <std::string>  phrases );
-void writeProtovisDataLevel( std::vector <std::string> tailPhrases, int curMultiplicativeVal, int numTabs );
+void writeProtovisDataLevel( std::vector <std::string> tailPhrases, long curMultiplicativeVal, int numTabs );
 
 int drawBar(float width);
 void translateYBy(int height);
